+++ rana_qgis_plugin/auth_3di.py
@@ -2,7 +2,7 @@ from qgis.core import QgsApplication, QgsAuthMethodConfig
 from qgis.PyQt.QtCore import QSettings
 
 from .communication import UICommunication
-from .constant import TENANT, THREEDI_AUTHCFG_ENTRY
+from .constant import THREEDI_AUTHCFG_ENTRY
 from .utils_api import get_threedi_personal_api_key, get_user_info
 
 
@@ -54,9 +54,8 @@ def setup_3di_auth(communication: UICommunication):
             return
     user = get_user_info(communication)
     if not user:
-        communication.show_error("Failed to get user information from Rana.")
         return
-    personal_api_key = get_threedi_personal_api_key(communication, TENANT, user["sub"])
+    personal_api_key = get_threedi_personal_api_key(communication, user["sub"])
     if personal_api_key:
         set_3di_auth(personal_api_key)
     else: