+++ rana_qgis_plugin/utils_api.py
@@ -4,10 +4,11 @@ from .auth import get_authcfg_id
 from .communication import UICommunication
 from .constant import API_URL, COGNITO_USER_INFO_ENDPOINT
 from .network_manager import NetworkManager
+from .utils import get_tenant_id
 
 
 class UserInfo(TypedDict):
-    sub: str
+    sub: str  # user_id
     given_name: str
     family_name: str
     email: str
@@ -21,30 +22,33 @@ def get_user_info(communication: UICommunication) -> Optional[UserInfo]:
     status, error = network_manager.fetch()
 
     if status:
-        user_info = network_manager.content
-        return user_info
+        user = network_manager.content
+        return user
     else:
         communication.show_error(f"Failed to get user info from cognito: {error}")
         return None
 
 
-def get_tenant(communication: UICommunication, tenant: str):
+def get_user_tenants(communication: UICommunication, user_id: str):
     authcfg_id = get_authcfg_id()
-    tenant_url = f"{API_URL}/tenants/{tenant}"
+    url = f"{API_URL}/tenants"
+    params = {"user_id": user_id}
 
-    network_manager = NetworkManager(tenant_url, authcfg_id)
-    status, error = network_manager.fetch()
+    network_manager = NetworkManager(url, authcfg_id)
+    status, error = network_manager.fetch(params)
 
     if status:
-        tenant = network_manager.content
-        return tenant
+        response = network_manager.content
+        items = response["items"]
+        return items
     else:
-        communication.show_error(f"Failed to get tenant: {error}")
-        return None
+        communication.show_error(f"Failed to get tenants: {error}")
+        return []
 
 
-def get_tenant_projects(communication: UICommunication, tenant: str):
+def get_tenant_projects(communication: UICommunication):
     authcfg_id = get_authcfg_id()
+    tenant = get_tenant_id()
     url = f"{API_URL}/tenants/{tenant}/projects"
 
     network_manager = NetworkManager(url, authcfg_id)
@@ -59,8 +63,9 @@ def get_tenant_projects(communication: UICommunication, tenant: str):
         return []
 
 
-def get_tenant_project_files(communication: UICommunication, tenant: str, project_id: str, params: dict = None):
+def get_tenant_project_files(communication: UICommunication, project_id: str, params: dict = None):
     authcfg_id = get_authcfg_id()
+    tenant = get_tenant_id()
     url = f"{API_URL}/tenants/{tenant}/projects/{project_id}/files/ls"
 
     network_manager = NetworkManager(url, authcfg_id)
@@ -75,8 +80,9 @@ def get_tenant_project_files(communication: UICommunication, tenant: str, projec
         return []
 
 
-def get_tenant_project_file(tenant: str, project_id: str, params: dict):
+def get_tenant_project_file(project_id: str, params: dict):
     authcfg_id = get_authcfg_id()
+    tenant = get_tenant_id()
     url = f"{API_URL}/tenants/{tenant}/projects/{project_id}/files/stat"
 
     network_manager = NetworkManager(url, authcfg_id)
@@ -89,8 +95,9 @@ def get_tenant_project_file(tenant: str, project_id: str, params: dict):
         return None
 
 
-def start_file_upload(tenant: str, project_id: str, params: dict):
+def start_file_upload(project_id: str, params: dict):
     authcfg_id = get_authcfg_id()
+    tenant = get_tenant_id()
     url = f"{API_URL}/tenants/{tenant}/projects/{project_id}/files/upload"
 
     network_manager = NetworkManager(url, authcfg_id)
@@ -103,8 +110,9 @@ def start_file_upload(tenant: str, project_id: str, params: dict):
         return None
 
 
-def finish_file_upload(tenant: str, project_id: str, payload: dict):
+def finish_file_upload(project_id: str, payload: dict):
     authcfg_id = get_authcfg_id()
+    tenant = get_tenant_id()
     url = f"{API_URL}/tenants/{tenant}/projects/{project_id}/files/upload"
     network_manager = NetworkManager(url, authcfg_id)
     status = network_manager.put(payload=payload)
@@ -114,8 +122,9 @@ def finish_file_upload(tenant: str, project_id: str, payload: dict):
     return None
 
 
-def get_threedi_schematisation(communication: UICommunication, tenant: str, descriptor_id: str):
+def get_threedi_schematisation(communication: UICommunication, descriptor_id: str):
     authcfg_id = get_authcfg_id()
+    tenant = get_tenant_id()
     url = f"{API_URL}/tenants/{tenant}/file-descriptors/{descriptor_id}/threedi-schematisation"
     network_manager = NetworkManager(url, authcfg_id)
     status, error = network_manager.fetch()
@@ -127,10 +136,11 @@ def get_threedi_schematisation(communication: UICommunication, tenant: str, desc
         return None
 
 
-def get_threedi_personal_api_key(communication: UICommunication, tenant: str, user_id: str) -> Optional[str]:
+def get_threedi_personal_api_key(communication: UICommunication, user_id: str) -> Optional[str]:
     communication.clear_message_bar()
     communication.bar_info("Getting 3Di personal API key ...")
     authcfg_id = get_authcfg_id()
+    tenant = get_tenant_id()
     url = f"{API_URL}/tenants/{tenant}/users/{user_id}/3di-personal-api-keys"
 
     network_manager = NetworkManager(url, authcfg_id)